refactor: show teams refactor
