+++ frontend/src/components/Teams/MyTeams/index.tsx
@@ -1,35 +1,19 @@
 import React from 'react';
-import { useQuery } from 'react-query';
-import { useSetRecoilState } from 'recoil';
 
-import { getTeamsOfUser } from 'api/teamService';
-import { toastState } from 'store/toast/atom/toast.atom';
-import { ToastStateEnum } from 'utils/enums/toast-types';
+import { Team } from '../../../types/team/team';
 import EmptyTeams from './partials/EmptyTeams';
 import ListOfCards from './partials/ListOfCards';
 
 type MyTeamsProps = {
 	userId: string;
+	teams: Team[];
+	isFetching: boolean;
 };
 
-const MyTeams = ({ userId }: MyTeamsProps) => {
-	const setToastState = useSetRecoilState(toastState);
+const MyTeams = ({ userId, teams, isFetching }: MyTeamsProps) => {
+	if (teams?.length === 0) return <EmptyTeams />;
 
-	const { data, isFetching } = useQuery(['teams'], () => getTeamsOfUser(), {
-		enabled: true,
-		refetchOnWindowFocus: false,
-		onError: () => {
-			setToastState({
-				open: true,
-				content: 'Error getting the teams',
-				type: ToastStateEnum.ERROR
-			});
-		}
-	});
-
-	if (!data) return <EmptyTeams />;
-
-	return <ListOfCards isLoading={isFetching} teams={data} userId={userId} />;
+	return <ListOfCards isLoading={isFetching} teams={teams} userId={userId} />;
 };
 
 export default MyTeams;
