+++ frontend/src/pages/teams/index.tsx
@@ -1,7 +1,8 @@
 import { ReactElement, Suspense } from 'react';
-import { dehydrate, QueryClient } from 'react-query';
+import { dehydrate, QueryClient, useQuery } from 'react-query';
 import { GetServerSideProps, GetServerSidePropsContext } from 'next';
 import { getSession, useSession } from 'next-auth/react';
+import { useSetRecoilState } from 'recoil';
 
 import { getDashboardHeaderInfo } from 'api/authService';
 import { getTeamsOfUser } from 'api/teamService';
@@ -10,16 +11,32 @@ import Layout from 'components/layouts/Layout';
 import LoadingPage from 'components/loadings/LoadingPage';
 import Flex from 'components/Primitives/Flex';
 import MyTeams from 'components/Teams/MyTeams';
+import { toastState } from '../../store/toast/atom/toast.atom';
+import { ToastStateEnum } from '../../utils/enums/toast-types';
 
 const Teams = () => {
 	const { data: session } = useSession({ required: true });
 
-	if (!session) return null;
+	const setToastState = useSetRecoilState(toastState);
+
+	const { data, isFetching } = useQuery(['teams'], () => getTeamsOfUser(), {
+		enabled: true,
+		refetchOnWindowFocus: false,
+		onError: () => {
+			setToastState({
+				open: true,
+				content: 'Error getting the teams',
+				type: ToastStateEnum.ERROR
+			});
+		}
+	});
+
+	if (!session || !data) return null;
 	return (
 		<Flex direction="column">
 			<Suspense fallback={<LoadingPage />}>
 				<QueryError>
-					<MyTeams userId={session.user.id} />
+					<MyTeams isFetching={isFetching} teams={data} userId={session.user.id} />
 				</QueryError>
 			</Suspense>
 		</Flex>
